Fixed the double->int misstake. Added PB test

+++ src/cpp_src/pybind_interface.cpp
@@ -52,7 +52,7 @@ PYBIND11_MODULE(nn_mwpc, m)
 nn_mwpc_interface::nn_mwpc_interface(const std::string& model_name, 
         int J_max_chn, double cutoff, int cut_pow, 
         bool sharp_cutoff, bool pre_comp_pot, bool rel_corr,
-        double number_of_p_points,bool finite_grid)
+        int number_of_p_points,bool finite_grid)
 {
 
     // ------ CONSTANTS TO CHANGE ------
